Convert Vehicle to Amenity
