+++ db/migrate/20230120010420_rename_vehicle_to_amenity.rb
@@ -1,0 +1,5 @@
+class RenameVehicleToAmenity < ActiveRecord::Migration[7.0]
+  def change
+    rename_table 'vehicles', 'amenities'
+  end
+end
